requirements.txt
# Core dependencies
requests>=2.31.0
Pillow>=10.0.0
pillow-heif>=0.13.0

# PDF generation
reportlab>=4.0.0

# Configuration and environment
python-dotenv>=1.0.0

# Optional: Advanced image processing
# rembg>=2.0.0  # AI background removal (requires onnxruntime)
# opencv-python>=4.8.0  # Advanced image processing

# Optional: Data management
# airtable-python-wrapper>=0.15.0  # Airtable integration

# Development
pytest>=7.4.0
black>=23.0.0
